Solution to Assignment2
I have writtrn code based on the example given that solves for cached
mean. The new code has the same logic. First it checks whether there are
already inverse cached. If yes, then it check whether the matrix we type
in is the same as the cached matrix. If again yes, then it return the
cached "m", inverse of the matrix. If no to either one, it means we want
to calculate inverse of a new matrix. The function will then cached the
matrix and its inverse.